backend/requirements.txt
django>=5.0,<6.0
djangorestframework
psycopg2-binary
django-cors-headers
